pyproject.toml
[project]
name = "spot2-challenge"
version = "0.1.0"
description = "Conversational AI for real estate assistant challenge."
authors = [
    {name = "Juan Vargas",email = "juan.vargasg2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "openai (>=1.76.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "fakeredis (>=2.28.1,<3.0.0)",
]

[tool.poetry]
packages = [{include = "spot2_challenge", from = "src"}]


[tool.poetry.group.dev.dependencies]
redis = "^5.2.1"
pydantic = {extras = ["dotenv"], version = "^2.11.3"}
pytest = "^8.3.5"
streamlit = "^1.44.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
